% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abbtostate}
\alias{abbtostate}
\title{Samantha Rhoads's function to...}
\usage{
abbtostate(v, state_names_abbs_df = get_states(return_which = "both"))
}
\description{
Samantha Rhoads's function to...
}
\examples{
abbtostate(v)
}
