% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_escape_characters}
\alias{add_escape_characters}
\title{Samantha Rhoads's function to add escape characters to a string}
\usage{
add_escape_characters(v)
}
\description{
Samantha Rhoads's function to add escape characters to a string
}
\examples{
add_escape_characters(v)
}
