% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_space_after_punct}
\alias{add_space_after_punct}
\title{Srhoads wrote this to allow you to...}
\usage{
add_space_after_punct(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
add_space_after_punct()
}
