% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_na}
\alias{all_na}
\title{A function to see whether all values are NA; can go inside `select_if()`}
\usage{
all_na(x)
}
\description{
A function to see whether all values are NA; can go inside `select_if()`
}
\examples{
all_na(x)
}
