% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{alnum_period}
\alias{alnum_period}
\title{A function}
\usage{
alnum_period(x)

alnum_period(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
alnum_period()
alnum_period()
}
