% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arrange_by_na}
\alias{arrange_by_na}
\title{This function is to arrange a dataframe by fewest NAs first (at the top of the dataset/earliest rows).}
\usage{
arrange_by_na(d)
}
\description{
This function is to arrange a dataframe by fewest NAs first (at the top of the dataset/earliest rows).
}
\examples{
arrange_by_na(d)
}
