% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arrange_by_na_desc}
\alias{arrange_by_na_desc}
\title{Function to arrange a dataframe by MOST NAs first (at the top of the dataset/earliest rows).}
\usage{
arrange_by_na_desc(d)
}
\description{
Function to arrange a dataframe by MOST NAs first (at the top of the dataset/earliest rows).
}
\examples{
{D <- data.frame(id=1:5, gender=c("F", "M", NA, "F", NA), race=c(3, 5, NA, 7, NA))}; arrange_by_na_desc(D); 
arrange_by_na_desc(d)
}
