% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.date.varioustypes}
\alias{as.date.varioustypes}
\title{A function to manipulate a string or vector into date format}
\usage{
as.date.varioustypes(v, include_time = F)
}
\description{
A function to manipulate a string or vector into date format
}
\examples{
as.date.varioustypes(v, include_time=F)
}
