% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as_numeric}
\alias{as_numeric}
\title{This function is same as as.numeric or readr::parse_number but u can do it on more than just a character vector + u can suppress the annoying warnings}
\usage{
as_numeric(v, suppresswarnings = T)
}
\description{
This function is same as as.numeric or readr::parse_number but u can do it on more than just a character vector + u can suppress the annoying warnings
}
\examples{
as_numeric(v, suppresswarnings=T)
}
