% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as_table_paste}
\alias{as_table_paste}
\title{Samantha Rhoads's function to copy something that looks like a dataframe (formatted like a dataframe) from a webpage/website & paste it in this function as one string & get a dataframe object back as output}
\usage{
as_table_paste(pastedstuff, sep = "\\t", header = T)
}
\description{
Srhoads wrote this to allow you to copy something that looks like a dataframe (formatted like a dataframe) from a webpage/website & paste it in this function as one string & get a dataframe object back as output
}
\examples{
as_table_paste()
}
