% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_range}
\alias{bin_range}
\title{Samantha Rhoads's function to take a numeric range and split/bin it by the number of groups you want}
\usage{
bin_range(v, n_groups = 10)
}
\description{
Samantha Rhoads's function to take a numeric range and split/bin it by the number of groups you want
}
\examples{
bin_range(v, n_groups=10)
}
