% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bind_rows2}
\alias{bind_rows2}
\title{Samantha Rhoads's function to bind rows going multiple depths}
\usage{
bind_rows2(l)
}
\description{
Samantha Rhoads's function to bind rows going multiple depths
}
\examples{
bind_rows2(l)
}
