% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bind_rows_}
\alias{bind_rows_}
\title{Samantha Rhoads's function to bind rows of object if they're dataframes}
\usage{
bind_rows_(..., .id = NULL)
}
\description{
Samantha Rhoads's function to bind rows of object if they're dataframes
}
\examples{
bind_rows_(..., .id = NULL)
}
