% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catn}
\alias{catn}
\title{This function is like `cat` but is wrapped in `\\n` (line breaks). It prints with formatting, and without quotes. Look at `cat`'s documentation for details}
\usage{
catn(
  ...,
  file = "",
  sep = " ",
  fill = F,
  labels = NULL,
  append = F,
  collapse = " "
)
}
\description{
This function is like `cat` but is wrapped in `\\n` (line breaks). It prints with formatting, and without quotes. Look at `cat`'s documentation for details
}
\examples{
catn(d) 
}
