% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_dfs}
\alias{clean_dfs}
\title{A clean_dfs Function}
\usage{
clean_dfs(df)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to
}
\examples{
clean_dfs()
}
\keyword{cats}
