% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_job_group}
\alias{clean_job_group}
\title{Samantha Rhoads's function to preliminarily clean a vector of job groups (jlcentric)}
\usage{
clean_job_group(v)
}
\description{
Samantha Rhoads's function to preliminarily clean a vector of job groups (jlcentric)
}
\examples{
clean_job_group(v)
}
