% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_jobtitle}
\alias{clean_jobtitle}
\title{Samantha Rhoads's function to clean a vector of job titles (jlcentric)}
\usage{
clean_jobtitle(v)
}
\description{
Samantha Rhoads's function to clean a vector of job titles (jlcentric)
}
\examples{
clean_jobtitle(v)
}
