% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_money_as_numeric}
\alias{clean_money_as_numeric}
\title{Srhoads wrote this to allow you to...}
\usage{
clean_money_as_numeric(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
clean_money_as_numeric()
}
