% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_punct}
\alias{clean_punct}
\title{Srhoads wrote this to allow you to clean punctuation in a string or vector of strings}
\usage{
clean_punct(v, removenegativesymbol = F, beginning = T, end = T)
}
\description{
Srhoads wrote this to allow you to clean punctuation in a string or vector of strings
}
\examples{
clean_punct()
}
