% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_recode}
\alias{clean_recode}
\title{A Function}
\usage{
clean_recode(df, scrub = c("once", "double"), extrarace = NULL,
  extragender = NULL)
}
\description{
This function allows you to
}
\examples{
clean_recode()
}
