% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{clean_str}
\alias{clean_str}
\title{A function}
\usage{
clean_str(v)

clean_str(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
clean_str()
clean_str()
}
