% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{cleanpath}
\alias{cleanpath}
\title{A Function}
\usage{
cleanpath(x)

cleanpath(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
cleanpath()
cleanpath()
}
