% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{col_types}
\alias{col_types}
\title{A function to get col_types when using `readr` package for functions like `read_csv()` with the argument `col_types=`}
\usage{
col_types(type = "c")
}
\description{
This function allows you to get col_types when using `readr` package for functions like `read_csv()` with the argument `col_types=`
}
\examples{
readr::read_csv("http://samplecsvs.s3.amazonaws.com/Sacramentorealestatetransactions.csv", col_types=col_types('c'))
col_types(type="c") 
}
