% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{combine_compact}
\alias{combine_compact}
\title{A Function}
\usage{
combine_compact(x)

combine_compact(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
combine_compact()
combine_compact()
}
