% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_two_lods}
\alias{compare_two_lods}
\title{Samantha Rhoads's function to compare the columns of two lists of dataframes to see their differences based on their `sumry()`s and if the column `sumry()`s are `identical()`}
\usage{
compare_two_lods(
  list1 = list(df = dfsampler()),
  list2 = list(df = dfsampler(which = "short")),
  exampleOfColumnName = "^name$"
)
}
\description{
Samantha Rhoads's function to compare the columns of two lists of dataframes to see their differences based on their `sumry()`s and if the column `sumry()`s are `identical()`
}
\examples{
compare_two_lods(list1 = list(df1 = dfsampler()), list2 =  list(df2 = dfsampler(which="short")), exampleOfColumnName="^name$")
}
