% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{connect_to_azure_file_share}
\alias{connect_to_azure_file_share}
\title{Samantha Rhoads's function to connect to an azure file share + return specific parameters or objects depending on what you want}
\usage{
connect_to_azure_file_share(
  share_name = "somestring",
  storage_name = "somestringalso",
  azure_key = "somelongstringugetfromazureportal",
  return_share = F
)
}
\description{
Samantha Rhoads's function to connect to an azure file share + return specific parameters or objects depending on what you want
}
\examples{
connect_to_azure_file_share(share_name="somestring", storage_name="somestringalso", azure_key="somelongstringugetfromazureportal", return_share=F)
}
