% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crosswalk_occp_codes}
\alias{crosswalk_occp_codes}
\title{Samantha Rhoads's function to crosswalk older OCCP (Census) Codes to the most current iteration (2018-2019) codes (jlcentric)}
\usage{
crosswalk_occp_codes(v, remove_comma_sep = F)
}
\description{
Samantha Rhoads's function to crosswalk older OCCP (Census) Codes to the most current iteration (2018-2019) codes (jlcentric)
}
\examples{
crosswalk_occp_codes(v, remove_comma_sep=F)
}
