% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{cvglmnet}
\alias{cvglmnet}
\title{A Function}
\usage{
cvglmnet(x = dtm_all, y = d$y, family = "binomial", alpha = 0,
  type.measure = "auc", nfolds = 5, thresh = 0.001, maxit = 1000)

cvglmnet(x = dtm_all, y = d$y, family = "binomial", alpha = 0,
  type.measure = "auc", nfolds = 5, thresh = 0.001, maxit = 1000)
}
\description{
This function allows you to

This function allows you to
}
\examples{
cvglmnet()
cvglmnet()
}
