% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_lol}
\alias{data_lol}
\title{A read files into list of list function}
\usage{
data_lol(pattern = "data", bindsheets = F, joinsheets = F)
}
\description{
This function allows you to read a list of files into a list of lists of data if the data is in excel (xlsx or xls format)
}
\examples{
data_lol(pattern='data', bindsheets=F, joinsheets=F)
}
