% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{datatable2}
\alias{datatable2}
\title{Srhoads wrote this to allow you to...}
\usage{
datatable2(
  x,
  vars = NULL,
  opts = NULL,
  caption = NULL,
  extensions = list(),
  ...
)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
datatable2(x, vars = NULL, opts = NULL, caption=NULL, extensions = list(), ...)
}
