% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{datatable_}
\alias{datatable_}
\title{This function allows you to make a DT datatable with better defaults in my opinion}
\usage{
datatable_(d)
}
\description{
This function allows you to make a DT datatable with better defaults in my opinion
}
\examples{
datatable_(d)
}
