% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dbDisconnectAll}
\alias{dbDisconnectAll}
\title{Samantha Rhoads's function to}
\usage{
dbDisconnectAll(drivername = "PostgreSQL")
}
\description{
Samantha Rhoads's function to
}
\examples{
dbDisconnectAll(drivername='PostgreSQL')
}
