% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dealwith_racegender_variable}
\alias{dealwith_racegender_variable}
\title{A Function}
\usage{
dealwith_racegender_variable(x)

dealwith_racegender_variable(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dealwith_racegender_variable()
()
}
