% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{describe_rgb_to_hex_translation}
\alias{describe_rgb_to_hex_translation}
\title{Samantha Rhoads's function to}
\usage{
describe_rgb_to_hex_translation(RGB = "rgb(35, 59, 150)")
}
\description{
Samantha Rhoads's function to
}
\examples{
describe_rgb_to_hex_translation(RGB="rgb(35, 59, 150)")
}
