% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df_get_preferred_column}
\alias{df_get_preferred_column}
\title{Samantha Rhoads's function to get relevant, preferred, desired column from dataframe on a loop}
\usage{
df_get_preferred_column(
  df,
  patterns = c("DateOpened", "Date.*Opened"),
  ignore.case = T,
  fillmissingwith = NA,
  returnNameOnly = F,
  exactEnd = F,
  exactStart = F,
<<<<<<< HEAD
  verbose = F,
  choose_non_empty_column = F,
  outcolname = "OUTCOLNAME"
=======
  verbose = F
>>>>>>> f405f1a7877bb7e35081dff4731c8a0360ba368e
)
}
\description{
Samantha Rhoads's function to get relevant, preferred, desired column from dataframe on a loop
}
\examples{
df_get_preferred_column(df, patterns=c('DateOpened', 'Date.*Opened'), ignore.case=T, fillmissingwith=NA, returnNameOnly=F, exactEnd=F, exactStart=F)
}
