% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{df_to_vec}
\alias{df_to_vec}
\title{A function}
\usage{
df_to_vec(df)

df_to_vec(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
df_to_vec()
df_to_vec()
}
