% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dfsampler}
\alias{dfsampler}
\title{A function that makes sample data frames with name, race, gender type columns. This function allows you to make sample dataframes!}
\usage{
dfsampler(which = c("long", "short")[1], tibble = F)
}
\description{
A function that makes sample data frames with name, race, gender type columns. This function allows you to make sample dataframes!
}
\examples{
dfsampler(which=c('long', 'short')[1], tibble=F)
}
