% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dimnna}
\alias{dimnna}
\title{A function}
\usage{
dimnna(d)

dimnna(d)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dimnna()
dimnna()
}
