% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{donothing}
\alias{donothing}
\title{A function to literally do nothing... just returns the identical object you fed it}
\usage{
donothing(x)
}
\description{
This function is to literally do nothing
}
\examples{
donothing(x)
}
