% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dply}
\alias{dply}
\title{A function to apply stuff to all columns in a df, specific to numeric or non-numeric type?}
\usage{
dply(df, fun, num = T)
}
\description{
A function to apply stuff to all columns in a df, specific to numeric or non-numeric type?
}
\examples{
dply()
}
