% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dply_all}
\alias{dply_all}
\title{A function}
\usage{
dply_all(x, fun)

dply_all(x, fun)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dply_all()
dply_all()
}
