% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dply_nonnumeric}
\alias{dply_nonnumeric}
\title{A function to apply stuff to just non-numeric columns in a df}
\usage{
dply_nonnumeric(x, fun)
}
\description{
A function to apply stuff to just non-numeric columns in a df
}
\examples{
dply_nonnumeric()
}
