% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dply_to_title}
\alias{dply_to_title}
\title{A function}
\usage{
dply_to_title(x, stringsAsFactors = T)

dply_to_title(x, stringsAsFactors = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dply_to_title()
dply_to_title()
}
