% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_empty}
\alias{drop_empty}
\title{A function to drop empty (NULL) items from a list}
\usage{
drop_empty(x_list)
}
\description{
A function to drop empty (NULL) items from a list
}
\examples{
drop_empty(x_list)
}
