% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_repeat_cols}
\alias{drop_repeat_cols}
\title{A function to paste a vector in a regex way with '|'}
\usage{
drop_repeat_cols(d, fromall = F, fromlast = F, fromfirst = F)
}
\description{
This function allows you to
}
\examples{
drop_repeat_cols(d, fromall=F, fromlast=F, fromfirst=F)
}
