% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_rows_n_cols_not_na}
\alias{drop_rows_n_cols_not_na}
\title{Samantha Rhoads's function to drop rows where not enough columns have data. IE: if you have 3 columns in a df but require at least 2 columns to have values, then you drop a row where two+ values are NA}
\usage{
drop_rows_n_cols_not_na(x, min_nonna_cols = 2)
}
\description{
Samantha Rhoads's function to drop rows where not enough columns have data. IE: if you have 3 columns in a df but require at least 2 columns to have values, then you drop a row where two+ values are NA
}
\examples{
drop_rows_n_cols_not_na(x, min_nonna_cols=2)
}
