% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dt_condensed}
\alias{dt_condensed}
\title{Srhoads wrote this to allow you to...}
\usage{
dt_condensed(df, pageLength = nrow(df))
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
dt_condensed()
}
