% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dt_datatables_pre}
\alias{dt_datatables_pre}
\title{Samantha Rhoads's function to...}
\usage{
dt_datatables_pre(df, pageLength = nrow(df))
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
dt_datatables_pre()
}
