% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DT_NAs_red_background}
\alias{DT_NAs_red_background}
\title{Samantha Rhoads's function to...}
\usage{
DT_NAs_red_background(DTdatatable)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
DT_NAs_red_background()
}
