% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dt_open}
\alias{dt_open}
\title{Samantha Rhoads's function to re-try running something multiple times (a designated number of `n_attempts`) if it fails until it succeeds.}
\usage{
dt_open(x, path = tempfile(fileext = ".html"))
}
\description{
Samantha Rhoads's function to re-try running something multiple times (a designated number of `n_attempts`) if it fails until it succeeds.
}
\examples{
retry_fxn(x, path=tempfile(fileext=".html"))
}
