% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dtmfunc}
\alias{dtmfunc}
\title{A Function}
\usage{
dtmfunc(df = d, corp = df$x, id = df$id, term_count_min = 2)

dtmfunc(df = d, corp = df$x, id = df$id, term_count_min = 2)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dtmfunc()
dtmfunc()
}
