% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edit_client_name}
\alias{edit_client_name}
\title{Samantha Rhoads's function to change one record of one variable in a SQL database (jlcentric)}
\usage{
edit_client_name(
  con,
  table = "clientmanagement_client",
  CLIENT_NAME_TO_RENAME = "GROSSOLDNAME",
  CLIENT_NAME_NEW = "PLEASEWORKCOOLNAME"
)
}
\description{
Samantha Rhoads's function to change one record of one variable in a SQL database (jlcentric)
}
\examples{
edit_client_name(con, table='clientmanagement_client', CLIENT_NAME_TO_RENAME = "GROSSOLDNAME", CLIENT_NAME_NEW = "PLEASEWORKCOOLNAME")
}
