% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excelToDateIf5DigitStr}
\alias{excelToDateIf5DigitStr}
\title{Samantha Rhoads's function to take an Excel 5-digit date and turn it into a date format}
\usage{
excelToDateIf5DigitStr(v)
}
\description{
Samantha Rhoads's function to take an Excel 5-digit date and turn it into a date format
}
\examples{
excelToDateIf5DigitStr(v)
}
