% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excelToDateIf5DigitStrAndManyDigitTime}
\alias{excelToDateIf5DigitStrAndManyDigitTime}
\title{Samantha Rhoads's function to...}
\usage{
excelToDateIf5DigitStrAndManyDigitTime(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
excelToDateIf5DigitStrAndManyDigitTime(v)
}
