% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_date}
\alias{extract_date}
\title{Samantha Rhoads's function to extract the date from a string, but doesn't turn it into a date!}
\usage{
extract_date(v)
}
\description{
Samantha Rhoads's function to extract the date from a string, but doesn't turn it into a date!
}
\examples{
extract_date()
}
