% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_digits}
\alias{extract_digits}
\title{Samantha Rhoads's function to extract numeric characters from a string}
\usage{
extract_digits(x)
}
\description{
Samantha Rhoads's function to extract numeric characters from a string
}
\examples{
extract_digits(x)
}
