% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_fourDigitYear}
\alias{extract_fourDigitYear}
\title{Samantha Rhoads's function to extract a year of 4 digits. Can be in the 1900s or 2000s only, or can include centuries before that}
\usage{
extract_fourDigitYear(string, includePre1900s = T)
}
\description{
Samantha Rhoads's function to extract a year of 4 digits. Can be in the 1900s or 2000s only, or can include centuries before that
}
\examples{
extract_fourDigitYear(string, includePre1900s=T)
}
