% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_state}
\alias{extract_state}
\title{Samantha Rhoads's function to extract states from a vector}
\usage{
extract_state(
  v,
  keep_which = c("all", "first", "last")[1],
  return_input_if_no_match = T
)
}
\description{
Samantha Rhoads's function to extract states from a vector
}
\examples{
extract_state(v, keep_which=c('all', 'first', 'last')[1], return_input_if_no_match=T)
}
