% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_word_startswith_dollarsign}
\alias{extract_word_startswith_dollarsign}
\title{Samantha Rhoads's function to...}
\usage{
extract_word_startswith_dollarsign(v)
}
\description{
Samantha Rhoads's function to...
}
\examples{
extract_word_startswith_dollarsign(v)
}
