% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{extract_words_w_nums}
\alias{extract_words_w_nums}
\title{A function}
\usage{
extract_words_w_nums(v)

extract_words_w_nums(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
extract_words_w_nums()
extract_words_w_nums()
}
