% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_year}
\alias{extract_year}
\title{Samantha Rhoads's function to extract the 4-digit year from a string, and stores it as a character!}
\usage{
extract_year(v)
}
\description{
Samantha Rhoads's function to extract the 4-digit year from a string, and stores it as a character!
}
\examples{
extract_year(v)
}
