% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{featherdump_anomalies_filelist_fun}
\alias{featherdump_anomalies_filelist_fun}
\title{A Function}
\usage{
featherdump_anomalies_filelist_fun(filelist = NULL, inpath = NULL,
  pattern = NULL, newdir = NULL, subsets = 2, subsubsets = 2,
  extra = NULL, outpath = "~/", filename_prefix = "DEFAULTNAME")
}
\description{
This function allows you to
}
\examples{
()
}
