% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{featherdump_anomalies_fun}
\alias{featherdump_anomalies_fun}
\title{A Function}
\usage{
featherdump_anomalies_fun(filelist = NULL, inpath = NULL,
  pattern = "\\\\.", outpath = "~", newdir = NULL, subsets = 2,
  subsubsets = 2, path = "~/", filename_prefix = "anomalies")
}
\description{
This function allows you to
}
\examples{
()
}
