% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{featherdump_regulars_anomalies}
\alias{featherdump_regulars_anomalies}
\title{A function}
\usage{
featherdump_regulars_anomalies(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  newdir = NULL,
  subsets = 2,
  subsubsets = 2,
  extra = NULL,
  outpath = "~/",
  filename_prefix = "DEFAULTNAME"
)
}
\description{
This function allows you to
}
\examples{
()
}
