% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fillr_by_at}
\alias{fillr_by_at}
\title{Samantha Rhoads's function to fill a dataframe variable (at_var) by a name of an ID variable (by_var)}
\usage{
fillr_by_at(d, by_var = "occ_code", at_var = "occ_description")
}
\description{
Samantha Rhoads's function to fill a dataframe variable (at_var) by a name of an ID variable (by_var)
}
\examples{
fillr_by_at(d, by_var='occ_code', at_var='occ_description')
}
