% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_duplicated}
\alias{filter_duplicated}
\title{A function that can filter a dataframe so it includes all duplicates of your variable of choice}
\usage{
filter_duplicated(d, var = "id")
}
\description{
A function that can filter a dataframe so it includes all duplicates of your variable of choice
}
\examples{
filter_duplicated()
}
