% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{findme}
\alias{findme}
\title{A function that finds strings in a dataframe anywhere (like `grep()` but many strings)}
\usage{
findme(d = df00.5, g1 = "", g2 = "", g3 = "", g4 = "", g5 = "")

findme(d = df00.5, g1 = "", g2 = "", g3 = "", g4 = "", g5 = "")

findme(d = df00.5, g1 = "", g2 = "", g3 = "", g4 = "", g5 = "")
}
\description{
This is a function

This is a function

This is a function
}
\examples{
findme(d=df00.5, g1="", g2="", g3="", g4="", g5="")
findme(d=df00.5, g1="", g2="", g3="", g4="", g5="")
findme(d=df00.5, g1="", g2="", g3="", g4="", g5="")
}
