% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{firstname_nomiddle}
\alias{firstname_nomiddle}
\title{A Function}
\usage{
firstname_nomiddle(v)

firstname_nomiddle(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
firstname_nomiddle()
firstname_nomiddle()
}
