% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{firstname_postcomma}
\alias{firstname_postcomma}
\title{A Function}
\usage{
firstname_postcomma(x)

firstname_postcomma(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
firstname_postcomma()
firstname_postcomma()
}
