% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flatten_list_from_top}
\alias{flatten_list_from_top}
\title{Samantha Rhoads's function to unlist/flatten a list one time, but from the top, instead of the bottom (aka the opposite of `unlist()` or purrr's `flatten()`)}
\usage{
flatten_list_from_top(L)
}
\description{
Samantha Rhoads's function to unlist/flatten a list one time, but from the top, instead of the bottom (aka the opposite of `unlist()` or purrr's `flatten()`)
}
\examples{
flatten_list_from_top(L)
}
