% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{format_initial}
\alias{format_initial}
\title{A function}
\usage{
format_initial(v, upper = T)

format_initial(v, upper = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
format_initial()
format_initial()
}
