% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fuzzy_match_rank}
\alias{fuzzy_match_rank}
\title{Samantha Rhoads's function to fuzzy-match (agrepl) a string to a dataframe column, returning a dataframe of matches with whichever other variables exist in the input dataframe}
\usage{
fuzzy_match_rank(
  s = "Data Scientist",
  strictest_max_distance = 0,
  seqstep = 0.01,
  df,
  dfvartocompare = "job",
  stop_at_n_matches = Inf
)
}
\description{
Samantha Rhoads's function to fuzzy-match (agrepl) a string to a dataframe column, returning a dataframe of matches with whichever other variables exist in the input dataframe
}
\examples{
fuzzy_match_rank(s="Data Scientist", strictest_max_distance=0, seqstep=.01, df, dfvartocompare="job", stop_at_n_matches=Inf)
}
