% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_firstname}
\alias{gather_firstname}
\title{A function}
\usage{
gather_firstname(df)

gather_firstname(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_firstname()
gather_firstname()
}
