% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_gender}
\alias{gather_gender}
\title{A Function}
\usage{
gather_gender(df)

gather_gender(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_gender()
()
}
