% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_lastname}
\alias{gather_lastname}
\title{A Function}
\usage{
gather_lastname(df)

gather_lastname(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_lastname()
gather_lastname()
}
