% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_name}
\alias{gather_name}
\title{A function}
\usage{
gather_name(df)

gather_name(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_name()
()
}
