% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_name_namesplit}
\alias{gather_name_namesplit}
\title{This function allows you to}
\usage{
gather_name_namesplit(df)
}
\description{
This function allows you to
}
\examples{
gather_name_namesplit()
}
