% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_namesplit}
\alias{gather_namesplit}
\title{A Function}
\usage{
gather_namesplit(df)

gather_namesplit(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_namesplit()
gather_namesplit()
}
