% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_namesplit_clean}
\alias{gather_namesplit_clean}
\title{This function allows you to}
\usage{
gather_namesplit_clean(df)
}
\description{
This function allows you to
}
\examples{
gather_namesplit_clean(df)
}
