% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gather_race}
\alias{gather_race}
\title{A Function}
\usage{
gather_race(df)

gather_race(df)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gather_race()
()
}
