% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gbsum}
\alias{gbsum}
\title{This function allows you to group by a desired variable and summarize it by `n()` (total count/sum). Compatible with dataframes & vectors. Uses `matches()` logic.}
\usage{
gbsum(d, var = NULL)
}
\description{
This function allows you to group by a desired variable and summarize it by `n()` (total count/sum). Compatible with dataframes & vectors. Uses `matches()` logic.
}
\examples{
gbsum(d, var=NULL) 
}
