% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geocode_by_cell}
\alias{geocode_by_cell}
\title{Srhoads wrote this to allow you to...}
\usage{
geocode_by_cell(v, replacewith = NULL)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
{geocode_by_cell('1156 Susan Way Sunnyvale, CA 94087')}; {geocode_by_cell('Stanford')}
geocode_by_cell()
}
