% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_column_letter}
\alias{get_column_letter}
\title{Samantha Rhoads's function to get the column letter of a number}
\usage{
get_column_letter(v)
}
\description{
Samantha Rhoads's function to get the column letter of a number
}
\examples{
get_column_letter(v)
}
