% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_state_puma}
\alias{get_state_puma}
\title{Samantha Rhoads's function to make state-puma out of state and puma}
\usage{
get_state_puma(
  st = c("CA"),
  puma = c("00100"),
  return_na_if_no_match = T,
  puma_nchar = 5
)
}
\description{
Samantha Rhoads's function to make state-puma out of state and puma
}
\examples{
get_state_puma(st=c("CA"), puma=c("00100"), return_na_if_no_match=T, puma_nchar=5)
}
