% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_states}
\alias{get_states}
\title{Samantha Rhoads's function to return either a HOLISTIC dataframe of state names and their abbreviations or a vector of just states or just abbreviations}
\usage{
get_states(return_which = c("both", "name", "abb", "fips")[1])
}
\description{
Samantha Rhoads's function to return either a HOLISTIC dataframe of state names and their abbreviations or a vector of just states or just abbreviations
}
\examples{
get_states(return_which=c('both', 'name', 'abb')[1])
}
