% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMostRecentFiles}
\alias{getMostRecentFiles}
\title{Samantha Rhoads's function to return the most recent versions of files based on a given pattern of characters of its name}
\usage{
getMostRecentFiles(
  path = ".",
  desc = T,
  verbose = F,
  pattern = NULL,
  all.files = FALSE,
  full.names = T,
  recursive = T,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE
)
}
\description{
Srhoads wrote this to allow you to get the most recent versions of files based on a given pattern of characters of its name
}
\examples{
getMostRecentFiles()
}
