% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grep_}
\alias{grep_}
\title{A function to do grep() but pasting a vec instead of just a string}
\usage{
grep_(pattern, v, exact = F, ignore.case = F, value = F)
}
\description{
This function allows you to do grep() but pasting a vec instead of just a string
}
\examples{
grep_(pattern, v, exact=F, ignore.case=F, value=F)
}
