% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{grep_all_df}
\alias{grep_all_df}
\title{A function to find matching string anywhere in a dataframe}
\usage{
grep_all_df(pattern, df, colnames = F, exact = F, ignore.case = F,
  print = F)

grep_all_df(pattern, df, colnames = F, exact = F, ignore.case = F,
  print = F)
}
\description{
This function allows you to find matching string anywhere in a dataframe

This function allows you to find matching string anywhere in a dataframe
}
\examples{
grep_all_df(pattern, df, colnames=F, exact=F, ignore.case=F, print=F, cells_only=F, cells_only_discrete=F, rownums_only=F)
grep_all_df(pattern, df, colnames=F, exact=F, ignore.case=F, print=F)
}
