% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grep_all_df_colnames}
\alias{grep_all_df_colnames}
\title{A function to find matching string anywhere in a dataframe}
\usage{
grep_all_df_colnames(pattern, df, exact = F, ignore.case = F, print = F)
}
\description{
This function allows you to find matching string anywhere in a dataframe
}
\examples{
grep_all_df_colnames(pattern, df, exact=F, ignore.case=F, print=F)
}
