% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grep_ic}
\alias{grep_ic}
\title{Samantha Rhoads's function to grep() but ignoring case as default & pasting a vec instead of just a string}
\usage{
grep_ic(
  pattern,
  x,
  ignore.case = T,
  perl = F,
  value = F,
  fixed = F,
  useBytes = F,
  invert = F
)
}
\description{
Samantha Rhoads's function to grep() but ignoring case as default & pasting a vec instead of just a string
}
\examples{
grep_ic()
}
