% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grepl_ic}
\alias{grepl_ic}
\title{Samantha Rhoads's function to grepl() but ignoring case as default & pasting a vec instead of just a string}
\usage{
grepl_ic(
  pattern,
  x,
  ignore.case = T,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\description{
Samantha Rhoads's function to grepl() but ignoring case as default & pasting a vec instead of just a string
}
\examples{
grepl_ic()
}
