% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grepval}
\alias{grepval}
\title{Samantha Rhoads's function to...}
\usage{
grepval(
  pattern,
  x,
  ignore.case = F,
  perl = F,
  value = T,
  fixed = F,
  useBytes = F,
  invert = F
)
}
\description{
Samantha Rhoads's function to...
}
\examples{
grepval(pattern, x, ignore.case=F, perl=F, value = T, fixed=F, useBytes=F, invert=F)
}
