% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gsub_}
\alias{gsub_}
\title{A function to do gsub() but pasting a vec instead of just a string}
\usage{
gsub_(pattern, to, v, exact = F, ignore.case = F)
}
\description{
This function allows you to do gsub() but pasting a vec instead of just a string
}
\examples{
gsub_(pattern, to, v, exact=F, ignore.case=F)
}
