% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gsub_NSRHOADS}
\alias{gsub_NSRHOADS}
\title{a function gsubing stuff related to srhoads jl desktop comp path}
\usage{
gsub_NSRHOADS(x)
}
\description{
This function allows you to
}
\examples{
gsub_NSRHOADS()
}
