% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{gsub_ply}
\alias{gsub_ply}
\title{A Function}
\usage{
gsub_ply(from, to, x, ignore.case = T, num = T)

gsub_ply(from, to, x, ignore.case = T, num = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
gsub_ply()
gsub_ply()
}
