% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{ifelse_to_na_exact}
\alias{ifelse_to_na_exact}
\title{A function}
\usage{
ifelse_to_na_exact(v, na_vec)

ifelse_to_na_exact(v, na_vec)
}
\description{
This function allows you to

This function allows you to
}
\examples{
ifelse_to_na_exact()
ifelse_to_na_exact()
}
