% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install.packages_wrapper}
\alias{install.packages_wrapper}
\title{A function to load and/or install package first (wrapper with tryCatch)}
\usage{
install.packages_wrapper(package, dependencies = NA, githubrepo = NULL,
  repos = c("https://cloud.r-project.org", "http://owi.usgs.gov/R/"),
  type = getOption("pkgType"))
}
\description{
This function allows you to load and/or install package first (wrapper with tryCatch)
}
\examples{
install.packages_wrapper()
}
