% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.factorchar}
\alias{is.factorchar}
\title{A function to ask if it's a character or factor logical}
\usage{
is.factorchar(x)
}
\description{
A function to ask if it's a character or factor logical
}
\examples{
is.factorchar(x)
}
