% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.nanull}
\alias{is.nanull}
\title{Samantha Rhoads's function to check if all of something is NA or NULL}
\usage{
is.nanull(x)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
is.nanull()
}
