% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.nanull_V1}
\alias{is.nanull_V1}
\title{Samantha Rhoads's function to check if all of something is NA or NULL}
\usage{
is.nanull_V1(x)
}
\description{
Srhoads wrote this to allow you to check if all of a variable is NA or NULL (old version from 20191210)
}
\examples{
is.nanull_V1()
}
