% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_datetype1}
\alias{is_datetype1}
\title{A function to determine if a string or vector is of date type}
\usage{
is_datetype1(v)
}
\description{
A function to determine if a string or vector is of date type
}
\examples{
is_datetype1(v)
}
