% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_local}
\alias{is_local}
\title{Samantha Rhoads's function to check if your in your local environment}
\usage{
is_local()
}
\description{
Samantha Rhoads's function to check if your in your local environment
}
\examples{
is_local()
}
