% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kill_special_space}
\alias{kill_special_space}
\title{Samantha Rhoads's function removes all special characters and spaces}
\usage{
kill_special_space(text)
}
\description{
Samantha Rhoads's function removes all special characters and spaces
}
\examples{
kill_special_space(text)
}
