% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{lapply2}
\alias{lapply2}
\title{A function}
\usage{
lapply2(l, fxn)

lapply2(l, fxn)

lapply2(l, fxn)
}
\description{
This function allows you to

This function allows you to

This function allows you to
}
\examples{
lapply2()
lapply2()
lapply2()
}
