% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lapply_on_nonumeric}
\alias{lapply_on_nonumeric}
\title{A function lapply stuff to just non-numeric columns in a df}
\usage{
lapply_on_nonumeric(df, fun)
}
\description{
A function lapply stuff to just non-numeric columns in a df
}
\examples{
lapply_on_nonumeric()
}
