% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{lastname_precomma}
\alias{lastname_precomma}
\title{A function}
\usage{
lastname_precomma(x)

lastname_precomma(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
lastname_precomma()
lastname_precomma()
}
