% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{lastname_wmiddle}
\alias{lastname_wmiddle}
\title{A function}
\usage{
lastname_wmiddle(v)

lastname_wmiddle(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
lastname_wmiddle()
lastname_wmiddle()
}
