% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_files}
\alias{list_files}
\title{same as list.files() except default recursive=T and full.names=T}
\usage{
list_files(p, pattern = NULL, recursive = T, full.names = T)
}
\description{
same as list.files() except default recursive=T and full.names=T
}
\examples{
list_files(p, pattern=NULL, recursive=T, full.names=T)
}
