% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{list_names}
\alias{list_names}
\title{A function}
\usage{
list_names(mylist, unique = F)

list_names(mylist, unique = F)
}
\description{
This function allows you to

This function allows you to see the names of inner list level in flat vec form
}
\examples{
list_names()
list_names()
}
