% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{list_to_df}
\alias{list_to_df}
\title{A function}
\usage{
list_to_df(mylist)

list_to_df(mylist)
}
\description{
This function allows you to

This function allows you to
}
\examples{
list_to_df()
()
}
