% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{load_rdata_files}
\alias{load_rdata_files}
\title{A function}
\usage{
load_rdata_files(files)

load_rdata_files(files)
}
\description{
This function allows you to

This function allows you to
}
\examples{
load_rdata_files()
load_rdata_files()
}
