% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookslike_number}
\alias{lookslike_number}
\title{A function}
\usage{
lookslike_number(
  v,
  include_decimal = F,
  include_comma = F,
  include_dash = F,
  include_space = F,
  include_all_punct = F
)
}
\description{
A function
}
\examples{
lookslike_number(v, include_decimal=F, include_comma=F, include_dash=F, include_space=F, include_all_punct=F)
}
