% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max_date}
\alias{max_date}
\title{Samantha Rhoads's function to find the max date appropriately, without getting Inf when there are NAs}
\usage{
max_date(dates, include_time = T, na.rm = T)
}
\description{
Samantha Rhoads's function to find the max date appropriately, without getting Inf when there are NAs
}
\examples{
max_date(dates, include_time=T, na.rm=T)
}
