% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{min_date}
\alias{min_date}
\title{Samantha Rhoads's function to find the min date appropriately, without getting Inf when there are NAs}
\usage{
min_date(dates, include_time = T, na.rm = T)
}
\description{
Samantha Rhoads's function to find the min date appropriately, without getting Inf when there are NAs
}
\examples{
min_date(dates, include_time=T, na.rm=T)
}
