% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{monthYear_to_dateRangeStr}
\alias{monthYear_to_dateRangeStr}
\title{Samantha Rhoads's function to}
\usage{
monthYear_to_dateRangeStr(
  v = c("2020-01, 2020-02, 2020-03", "2020-05, 2021-01",
    "2021-10, 2021-11, 2021-12, 2022-02"),
  to_symbol = " to "
)
}
\description{
Samantha Rhoads's function to
}
\examples{
monthYear_to_dateRangeStr(v = c('2020-01, 2020-02, 2020-03', '2020-05, 2021-01', '2021-10, 2021-11, 2021-12, 2022-02'), to_symbol=" to ")
}
