% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistep_clean_recode}
\alias{multistep_clean_recode}
\title{A function}
\usage{
multistep_clean_recode(
  df,
  subsets = 2,
  scrub = NULL,
  extrarace = NULL,
  extragender = NULL
)
}
\description{
This function allows you to
}
\examples{
()
}
