% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistep_gather_join_first_last_name}
\alias{multistep_gather_join_first_last_name}
\title{A function}
\usage{
multistep_gather_join_first_last_name(df, subsets = 2)
}
\description{
This function allows you to
}
\examples{
()
}
