% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistep_preprocess_all_cols}
\alias{multistep_preprocess_all_cols}
\title{A function}
\usage{
multistep_preprocess_all_cols(
  mylist,
  type = NULL,
  subsets = 2,
  subsubsets = 2,
  write = FALSE,
  featherpath = "~/"
)
}
\description{
This function allows you to
}
\examples{
multistep_preprocess_all_cols(mylist, type = NULL, subsets = 2, subsubsets = 2, write = FALSE, featherpath = "~/")
}
