% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistep_split_files}
\alias{multistep_split_files}
\title{A function}
\usage{
multistep_split_files(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  newdir = NULL,
  subsets = NULL,
  by = NULL,
  extra = NULL,
  outpath = "~/",
  filename_prefix = "DEFAULTNAME"
)
}
\description{
This function allows you to
}
\examples{
()
}
