% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_col_if_not_exists}
\alias{mutate_col_if_not_exists}
\title{Samantha Rhoads's function to create a column in a dataframe if it doesn't already exist; defaults to all NA values}
\usage{
mutate_col_if_not_exists(d, var = "YOS_Y")
}
\description{
Samantha Rhoads's function to create a column in a dataframe if it doesn't already exist; defaults to all NA values
}
\examples{
mutate_col_if_not_exists(d, var='id')
}
