% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_listname_to_lod_column}
\alias{mutate_listname_to_lod_column}
\title{Srhoads wrote this to allow you to...}
\usage{
mutate_listname_to_lod_column(LoD, new_colname_for_listname = "LISTNAME")
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
mutate_listname_to_lod_column(LoD, new_colname_for_listname="LISTNAME")
}
