% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{na_if_}
\alias{na_if_}
\title{A function}
\usage{
na_if_(x)

na_if_(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
na_if_()
na_if_()
}
