% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{na_if_all_digits}
\alias{na_if_all_digits}
\title{A function}
\usage{
na_if_all_digits(v)

na_if_all_digits(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
na_if_all_digits()
na_if_all_digits()
}
