% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{naics_to_sector}
\alias{naics_to_sector}
\title{Samantha Rhoads's function to convert NAICS codes to their sector names or codes (jlcentric)}
\usage{
naics_to_sector(v, return_description = F)
}
\description{
Samantha Rhoads's function to convert NAICS codes to their sector names or codes (jlcentric)
}
\examples{
naics_to_sector(v, return_description=F)
}
