% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{name_case}
\alias{name_case}
\title{A function}
\usage{
name_case(v)

name_case(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
name_case()
name_case()
}
