% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{newdate}
\alias{newdate}
\title{A function to clean dates: recode a date that's originally in a format like `January 17, 1996 11:00 AM`}
\usage{
newdate(v)

newdate(v)
}
\description{
A function to clean dates: recode a date that's originally in a format like `January 17, 1996 11:00 AM`

A function
}
\examples{
newdate()
newdate()
}
