% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not_all_same}
\alias{not_all_same}
\title{A function that can go inside `select_if()` and checks if all items in a vector are identical or not, returns TRUE if they're not all identical}
\usage{
not_all_same(x)
}
\description{
A function that can go inside `select_if()` and checks if all items in a vector are identical or not, returns TRUE if they're not all identical
}
\examples{
not_all_same()
}
