% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not_is.character}
\alias{not_is.character}
\title{Srhoads wrote this to allow you to...}
\usage{
not_is.character(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
not_is.character()
}
