% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nword}
\alias{nword}
\title{Samantha Rhoads's function to return the number of words in a string}
\usage{
nword(v, split_by_space_only = F)
}
\description{
Samantha Rhoads's function to return the number of words in a string
}
\examples{
nword(v, split_by_space_only=F)
}
