% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pad_leading_0s}
\alias{pad_leading_0s}
\title{Samantha Rhoads's function to pad 0s in front of string to make string a certain length}
\usage{
pad_leading_0s(
  v = c("1234"),
  length = 5,
  empty_if_all_0s = T,
  only_numbers = F
)
}
\description{
Samantha Rhoads's function to pad 0s in front of string to make string a certain length
}
\examples{
recode_0s_pad_str(v)
}
