% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_excel_date}
\alias{parse_excel_date}
\title{#' A function to parse an excel dates: this function allows you to parse an excel dates (one of those with 5 digits as a string)
#' @export
#' @examples
#' parse_excel_date(v, include_time=F)}
\usage{
parse_excel_date(v, include_time = F)
}
\description{
#' A function to parse an excel dates: this function allows you to parse an excel dates (one of those with 5 digits as a string)
#' @export
#' @examples
#' parse_excel_date(v, include_time=F)
}
