% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_regex}
\alias{paste_regex}
\title{A function to paste a vector in a regex way with '|' options for exact or partial}
\usage{
paste_regex(v, collapse = "|", exact = F)
}
\description{
This function allows you to paste a vector in a regex way with '|' options for exact or partial
}
\examples{
paste_regex()
}
