% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_regex_partial}
\alias{paste_regex_partial}
\title{A function to paste a vector in a regex way with '|' partial}
\usage{
paste_regex_partial(v, collapse = "|")

paste_regex_partial(v, collapse = "|")
}
\description{
This function allows you to paste a vector in a regex way with '|' partial

This function allows you to
}
\examples{
paste_regex_partial(v, collapse='|')
paste_regex_partial()
}
