% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_unique_sep_sort}
\alias{paste_unique_sep_sort}
\title{Samantha Rhoads's function to}
\usage{
paste_unique_sep_sort(v, sep = "; ", collapse = sep, sort_str = T)
}
\description{
Samantha Rhoads's function to
}
\examples{
paste_unique_sep_sort(v, sep = "; ", collapse=sep)
}
