% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pkg4}
\alias{pkg4}
\title{A function to load and/or install package first (holistic)!}
\usage{
pkg4(package1 = NULL, ..., pipes = F, dependencies = NA,
  githubrepo = NULL, repos = c("https://cloud.r-project.org",
  "http://owi.usgs.gov/R/"), type = getOption("pkgType"))
}
\description{
This function allows you to load and/or install package first (holistic)!
}
\examples{
pkg4()
}
