% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pkg_tidy}
\alias{pkg_tidy}
\title{This function allows you to load and/or install package first (tidyverse included by default)!}
\usage{
pkg_tidy(package1 = "tidyverse", ...)
}
\description{
This function allows you to load and/or install package first (tidyverse included by default)!
}
\examples{
pkg_tidy(package1="tidyverse", ...)
}
