% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{ply}
\alias{ply}
\title{A Function}
\usage{
ply(x, fun, num = T)

ply(x, fun, num = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
ply()
ply()
}
