% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prestep_preprocess_data}
\alias{prestep_preprocess_data}
\title{A Function}
\usage{
prestep_preprocess_data(mylist, subsets = 2, type = NULL,
  extent = NULL)
}
\description{
This function allows you to
}
\examples{
()
}
