% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preview_hex_colors}
\alias{preview_hex_colors}
\title{Samantha Rhoads's function to visualize hex color codes}
\usage{
preview_hex_colors(
  hexs = c("#b41e3b", "#337ab7", "#f37036", "#434447", "#37718e", "#da3051", "#f5f6f8",
    "#9a2138", "#ffffff", "#333333"),
  plotly = F
)
}
\description{
Samantha Rhoads's function to visualize hex color codes
}
\examples{
preview_hex_colors(hexs = c("#b41e3b", "#337ab7", "#f37036", "#434447", "#37718e", "#da3051", "#f5f6f8"))
}
