% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rand_str}
\alias{rand_str}
\title{Samantha Rhoads's function to generate a random string}
\usage{
rand_str(n = 1, length = 10, pattern = "[a-zA-Z0-9]")
}
\description{
Samantha Rhoads's function to generate a random string
}
\examples{
rand_str()
}
