% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_csvs}
\alias{read_csvs}
\title{A function to read csv files}
\usage{
read_csvs(filelist, bindrows = F, simplif = T)

read_csvs(filelist, bindrows = F, simplif = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
read_csvs()
read_csvs()
}
