% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_df_all}
\alias{read_df_all}
\title{A Function}
\usage{
read_df_all(filenames, bindrows = F, bindsheets = F, simplif = T)

read_df_all(filenames, bindrows = F, bindsheets = F, simplif = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
read_df_all()
read_df_all()
}
