% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dfs_process_by1}
\alias{read_dfs_process_by1}
\title{A Function}
\usage{
read_dfs_process_by1(filelist, outpath = "AA/data/", prefix = "auto",
  startdoc = 1)
}
\description{
This function allows you to
}
\examples{
()
}
