% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_excels}
\alias{read_excels}
\title{A function to read excel files}
\usage{
read_excels(filelist, bindsheets = F, bindrows = F, simplif = T,
  col_types = "text")

read_excels(filelist, bindsheets = F, bindrows = F, simplif = T,
  col_types = "text")
}
\description{
This function allows you to

This function allows you to
}
\examples{
read_excels()
read_excels()
}
