% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_extract_merge}
\alias{read_extract_merge}
\title{A Function}
\usage{
read_extract_merge(filelist = NULL, inpath = NULL, pattern = NULL,
  Ng = F, Nr = F, Nrg = F)
}
\description{
This function allows you to
}
\examples{
()
}
