% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_feathers}
\alias{read_feathers}
\title{A function}
\usage{
read_feathers(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  bind = T,
  recode_extra_na = F
)
}
\description{
This function allows you to
}
\examples{
read_feathers()
}
