% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_feathers_recode_write}
\alias{read_feathers_recode_write}
\title{A Function}
\usage{
read_feathers_recode_write(filelist = NULL, inpath = NULL,
  pattern = NULL, outpath = "", newdir = NULL)
}
\description{
This function allows you to
}
\examples{
()
}
