% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files}
\alias{read_files}
\title{A function to read multiple of any files from read_file()}
\usage{
read_files(filelist, bindsheets = F, joinsheets = F, bindrows = F,
  simplif = T)
}
\description{
This function allows you to
}
\examples{
read_files(filelist, bindsheets=F,joinsheets=F, bindrows=F, simplif=T)
}
