% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files_allsheets}
\alias{read_files_allsheets}
\title{A function to read multiple Excel or CSV (including XLSB) files and either all or selective sheets from them.}
\usage{
read_files_allsheets(
  fns = NULL,
  keepshtvec = NULL,
  na = c("NA", "None", "N/A", "-", ""),
  col_types = "text",
  skip = 0,
  col_names = T,
  range = NULL,
  trim_ws = T,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl::readxl_progress(),
  .name_repair = "unique",
  xlsb_sheets = NULL
)
}
\description{
A function to read multiple Excel or CSV (including XLSB) files and either all or selective sheets from them.
}
\examples{
read_files_allsheets(fns=NULL, keepshtvec=NULL, na=c("NA", "None", "N/A", "-", ""), col_types="text", skip=0, col_names=T, range=NULL, trim_ws=T, n_max=Inf, guess_max=min(1000, n_max), progress=readxl::readxl_progress(), .name_repair="unique")
}
