% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files_to_feather}
\alias{read_files_to_feather}
\title{A function}
\usage{
read_files_to_feather(filelist, path = "~/", filename_prefix = "nrg_data")
}
\description{
A function
}
\examples{
read_files_to_feather()
}
