% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files_to_rda}
\alias{read_files_to_rda}
\title{A Function}
\usage{
read_files_to_rda(filelist, path = "~/", filename_prefix = "nrg_data",
  col_types = "text")
}
\description{
This function allows you to
}
\examples{
()
}
