% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_merge_write_feathers}
\alias{read_merge_write_feathers}
\title{A function}
\usage{
read_merge_write_feathers(filelist = NULL, inpath = NULL,
  pattern = NULL, newdir = NULL, outpath = "~/",
  filename_prefix = "DEFAULT")

read_merge_write_feathers(filelist = NULL, inpath = NULL,
  pattern = NULL, newdir = NULL, outpath = "~/",
  filename_prefix = "DEFAULT")

read_merge_write_feathers(filelist = NULL, inpath = NULL,
  pattern = NULL, newdir = NULL, outpath = "~/",
  filename_prefix = "DEFAULT")
}
\description{
This function allows you to

This function allows you to

This function allows you to
}
\examples{
read_merge_write_feathers()
()
read_merge_write_feathers()
}
