% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_rdas}
\alias{read_rdas}
\title{A function}
\usage{
read_rdas(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  bind = F,
  recode_extra_na = F
)

read_rdas(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  bind = F,
  recode_extra_na = F
)
}
\description{
This function allows you to

This function allows you to
}
\examples{
read_rdas()
read_rdas()
}
