% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_ydrive_clean_write}
\alias{read_ydrive_clean_write}
\title{A function}
\usage{
read_ydrive_clean_write(
  filenames,
  csv = F,
  xlsx = F,
  xls = F,
  outpath = "data/original/",
  col_types = "text"
)
}
\description{
This function allows you to
}
\examples{
read_ydrive_clean_write()
}
