% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readexcel}
\alias{readexcel}
\title{A function just like read_excel but better!}
\usage{
readexcel(file, bindsheets = F, joinsheets = F)
}
\description{
This function allows you to read an xls or xlsx efficiently, even w/ multiple sheets!
}
\examples{
readexcel(file, bindsheets=F, joinsheets=F)
}
