% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_date}
\alias{recode_date}
\title{Samantha Rhoads's function to force recode something that looks like it might be a date into a more consistent date structure.}
\usage{
recode_date(v)
}
\description{
Samantha Rhoads's function to force recode something that looks like it might be a date into a more consistent date structure.
}
\examples{
retry_fxn(x, path=tempfile(fileext=".html"))
}
