% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_diversitygroup}
\alias{recode_diversitygroup}
\title{Srhoads wrote this to allow you to...}
\usage{
recode_diversitygroup(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
recode_diversitygroup()
}
