% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_eeo1}
\alias{recode_eeo1}
\title{Samantha Rhoads's function to recode eeo1 category with broader reaches than recode_jobgroup()... (jlcentric)}
\usage{
recode_eeo1(v, return_codes = F)
}
\description{
Samantha Rhoads's function to recode eeo1 category with broader reaches than recode_jobgroup()... (jlcentric)
}
\examples{
recode_eeo1(v, return_codes=F)
}
