% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_gender}
\alias{recode_gender}
\title{Samantha Rhoads's function to recode gender more quickly than prior versions!!}
\usage{
recode_gender(v)
}
\description{
Samantha Rhoads's function to recode gender more quickly than prior versions!!
}
\examples{
recode_gender(v)
}
