% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_gender2}
\alias{recode_gender2}
\title{A function}
\usage{
recode_gender2(vec, recode_list = gender_list, extra = NULL)

recode_gender2(vec, recode_list = gender_list, extra = NULL)
}
\arguments{
\item{recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to

This function allows you to
}
\examples{
recode_gender2()
recode_gender2()
}
