% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_gender_j}
\alias{recode_gender_j}
\title{A Recode Gender Function}
\usage{
recode_gender_j(
  vec,
  recode_list = gender_list,
  extra = c("nothing", "other", "NA")
)

recode_gender_j(
  vec,
  recode_list = gender_list,
  extra = c("nothing", "other", "NA")
)
}
\arguments{
\item{recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to recode gender!

This function allows you to recode gender!
}
\examples{
recode_gender_j()
recode_gender_j()
}
\keyword{cats}
