% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_gender_regex}
\alias{recode_gender_regex}
\title{A function to recode gender!}
\usage{
recode_gender_regex(vec)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
A function to recode gender!
}
\examples{
recode_gender_regex()
}
\keyword{cats}
