% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_gender_specific}
\alias{recode_gender_specific}
\title{A Function}
\usage{
recode_gender_specific(vec, extra = NULL)

recode_gender_specific(vec, extra = NULL)
}
\description{
This function allows you to

This function allows you to
}
\examples{
recode_gender_specific()
recode_gender_specific()
}
