% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_genders_regex}
\alias{recode_genders_regex}
\title{This function allows you to recode gender!}
\usage{
recode_genders_regex(df)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to recode gender!
}
\examples{
recode_genders_regex()
}
\keyword{cats}
