% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_jobgroup}
\alias{recode_jobgroup}
\title{Samantha Rhoads's function to recode EEO1 category (job group) either with component 2 logic or regular logic (jlcentric)}
\usage{
recode_jobgroup(v, component2 = F, verbose = F)
}
\description{
Samantha Rhoads's function to recode EEO1 category (job group) either with component 2 logic or regular logic (jlcentric)
}
\examples{
recode_jobgroup(v, component2=F, verbose=F)
}
