% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_jobgroup_nums}
\alias{recode_jobgroup_nums}
\title{Samantha Rhoads's function to recode jobgroups as numbers rounded to the nearest tenth (ie: 1.1 or 1.2) if the string looks like a number (jlcentric)}
\usage{
recode_jobgroup_nums(v, verbose = F)
}
\description{
Samantha Rhoads's function to recode jobgroups as numbers rounded to the nearest tenth (ie: 1.1 or 1.2) if the string looks like a number (jlcentric)
}
\examples{
recode_jobgroup_nums(v, verbose=F)
}
