% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_lag_while_NAs}
\alias{recode_lag_while_NAs}
\title{Samantha Rhoads's function to fill in missing values with the value directly above the given cell with an NA in a specific column of a dataframe}
\usage{
recode_lag_while_NAs(x, colname = "eeid")
}
\description{
Samantha Rhoads's function to fill in missing values with the value directly above the given cell with an NA in a specific column of a dataframe
}
\examples{
recode_lag_while_NAs(x, colname="eeid")
}
