% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_match}
\alias{recode_match}
\title{Samantha Rhoads's function to recode a vector from 2 reference vectors}
\usage{
recode_match(
  v_input = c("girl", "dude"),
  match_vec = c("dude", "man", "girl", "woman"),
  desired_vec = c("male", "male", "female", "female")
)
}
\description{
Samantha Rhoads's function to recode a vector from 2 reference vectors
}
\examples{
recode_match(v_input=c('girl', 'dude'), match_vec=c('dude', 'man', 'girl', 'woman'), desired_vec=c('male', 'male', 'female', 'female'))
}
