% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_minority}
\alias{recode_minority}
\title{Samantha Rhoads's function to recode specific races to minorities}
\usage{
recode_minority(v = "subrace")
}
\description{
Samantha Rhoads's function to recode specific races to minorities
}
\examples{
recode_minority(v="subrace")
}
