% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_na_old}
\alias{recode_na_old}
\title{A function to recode specified values into NAs}
\usage{
recode_na_old(x)
}
\description{
A function to recode specified values into NAs
}
\examples{
recode_na_old(x)
}
