% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_na_vec}
\alias{recode_na_vec}
\title{A function}
\usage{
recode_na_vec(vec, recode_list = recode_na_list, extra = NULL)

recode_na_vec(vec, recode_list = recode_na_list, extra = NULL)
}
\description{
This function allows you to

This function allows you to
}
\examples{
recode_na_vec()
()
}
