% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_path}
\alias{recode_path}
\title{Samantha Rhoads's function to convert a Windows-readable-path to a Mac-readable path, or vice-versa}
\usage{
recode_path(path)
}
\description{
Samantha Rhoads's function to convert a Windows-readable-path to a Mac-readable path, or vice-versa
}
\examples{
recode_path(path)
}
