% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_race}
\alias{recode_race}
\title{Samantha Rhoads's function to recode race more quickly than prior versions!!}
\usage{
recode_race(v)
}
\description{
Samantha Rhoads's function to recode race more quickly than prior versions!!
}
\examples{
recode_race(v)
}
