% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_race2}
\alias{recode_race2}
\title{A function}
\usage{
recode_race2(vec, recode_list = race_list, extra = NULL)
}
\description{
This function allows you to
}
\examples{
recode_race2(vec, recode_list = race_list, extra = NULL)
}
