% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_race_j}
\alias{recode_race_j}
\title{A Function}
\usage{
recode_race_j(vec, recode_list = race_list, extra = c("nothing",
  "multirace", "NA"))
}
\description{
This function allows you to
}
\examples{
()
}
