% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_race_or_gender}
\alias{recode_race_or_gender}
\title{A Recode Race or Gender Function}
\usage{
recode_race_or_gender(
  vec,
  recode_list = list(`NA` = c("undef", "none", "unknown"))
)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
A Recode Race or Gender Function
}
\examples{
recode_race_or_gender()
}
\keyword{cats}
