% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_race_regex}
\alias{recode_race_regex}
\title{A Function}
\usage{
recode_race_regex(vec)

recode_race_regex(vec)
}
\description{
This function allows you to

This function allows you to
}
\examples{
recode_race_regex()
recode_race_regex()
}
