% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_races_and_genders}
\alias{recode_races_and_genders}
\title{A Function}
\usage{
recode_races_and_genders(df, extragender = NULL, extrarace = NULL,
  extra = NULL)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to
}
\examples{
recode_races_and_genders()
}
\keyword{cats}
