% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_races_and_genders_regex}
\alias{recode_races_and_genders_regex}
\title{A function}
\usage{
recode_races_and_genders_regex(df, extragender = NULL,
  extrarace = NULL)

recode_races_and_genders_regex(df, extragender = NULL,
  extrarace = NULL)
}
\arguments{
\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}

\item{#recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{#extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to recode gender!

This function allows you to recode gender!
}
\examples{
recode_races_and_genders_regex()
recode_races_and_genders_regex()
}
\keyword{cats}
