% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{reduce_name_bytes}
\alias{reduce_name_bytes}
\title{A function}
\usage{
reduce_name_bytes(x)

reduce_name_bytes(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
reduce_name_bytes()
reduce_name_bytes()
}
