% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regex_info}
\alias{regex_info}
\title{A function that reminds us what some regular expressions are, like and/or}
\usage{
regex_info()
}
\description{
A function that reminds us what some regular expressions are, like and/or
}
\examples{
regex_info()
}
