% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regulars_namesplit}
\alias{regulars_namesplit}
\title{A Function}
\usage{
regulars_namesplit(x, extra = NULL, extent = "thorough")
}
\description{
This function allows you to
}
\examples{
regulars_namesplit()
}
