% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_dir}
\alias{remove_dir}
\title{Samantha Rhoads's function to delete a directory and all of its contents}
\usage{
remove_dir(path)
}
\description{
Samantha Rhoads's function to delete a directory and all of its contents
}
\examples{
remove_dir(path)
}
