% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_duplicate_punct}
\alias{remove_duplicate_punct}
\title{Samantha Rhoads's function to...}
\usage{
remove_duplicate_punct(v, nrepeat = 6)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
remove_duplicate_punct()
}
