% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_duplicate_punct_separated_by_space}
\alias{remove_duplicate_punct_separated_by_space}
\title{Srhoads wrote this to allow you to...}
\usage{
remove_duplicate_punct_separated_by_space(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
remove_duplicate_punct_separated_by_space()
}
