% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_space_btwn_identical_punct}
\alias{remove_space_btwn_identical_punct}
\title{Samantha Rhoads's function to...}
\usage{
remove_space_btwn_identical_punct(v)

remove_space_btwn_identical_punct(v)
}
\description{
Srhoads wrote this to allow you to...

Srhoads wrote this to allow you to...
}
\examples{
remove_space_btwn_identical_punct()
remove_space_btwn_identical_punct()
}
