% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reticulate_correctly}
\alias{reticulate_correctly}
\title{Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from}
\usage{
reticulate_correctly()
}
\description{
Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from
}
\examples{
reticulate_correctly(x, y, printWhichOnly=F)
}
