% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retry_fxn}
\alias{retry_fxn}
\title{Samantha Rhoads's function to re-try running something multiple times (a designated number of `n_attempts`) if it fails until it succeeds.}
\usage{
retry_fxn(FUN = function() {
    "hi"
}, n_attempts = 5)
}
\description{
Samantha Rhoads's function to re-try running something multiple times (a designated number of `n_attempts`) if it fails until it succeeds.
}
\examples{
retry_fxn(FUN=function(){"hi"}, n_attempts=5)
}
