% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{returnname}
\alias{returnname}
\title{A function}
\usage{
returnname(z)

returnname(z)
}
\description{
This function allows you to

This function allows you to
}
\examples{
returnname()
returnname()
}
