% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rm_ds_store}
\alias{rm_ds_store}
\title{This function is a system wrapper to delete silly .DS_Store files}
\usage{
rm_ds_store()
}
\description{
This function is a system wrapper to delete silly .DS_Store files
}
\examples{
rm_ds_store(v, suppresswarnings=T)
}
