% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{round5}
\alias{round5}
\title{A Function}
\usage{
round5(num)

round5(num)
}
\description{
This function allows you to

This function allows you to
}
\examples{
round5()
round5()
}
