% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round_intuitively}
\alias{round_intuitively}
\title{A function to round intuitively based on how many digits the percentage number is + how much space you want to take up}
\usage{
round_intuitively(x, roundby = 2)
}
\description{
A function to round intuitively based on how many digits the percentage number is + how much space you want to take up
}
\examples{
round_intuitively(x, roundby=2)
}
