% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round_nearest_minute}
\alias{round_nearest_minute}
\title{Samantha Rhoads's function to round a datetime to the nearest minute}
\usage{
round_nearest_minute(v)
}
\description{
Samantha Rhoads's function to round a datetime to the nearest minute
}
\examples{
round_nearest_minute(v)
}
