% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roundupc}
\alias{roundupc}
\title{A function that rounds up to next highest digit base #, ie, to 100 or 1000 or 1000 etc...}
\usage{
roundupc(x)
}
\description{
A function that rounds up to next highest digit base #, ie, to 100 or 1000 or 1000 etc...
}
\examples{
roundupc(x)
}
