% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{samtokens}
\alias{samtokens}
\title{A Function}
\usage{
samtokens(df, nmin = 2, nmax = 3, xv = "firstname", yv = "gender",
  sample = nrow(df))
}
\description{
This function allows you to
}
\examples{
()
}
