% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_at_if}
\alias{select_at_if}
\title{Samantha Rhoads's function to select columns from a dataframe based on colname match and data contents match}
\usage{
select_at_if(
  d,
  colname_pattern = vars(everything()),
  data_pattern = function(x) {
     T
 },
  condition = c("or", "and")[1]
)
}
\description{
Samantha Rhoads's function to select columns from a dataframe based on colname match and data contents match
}
\examples{
select_at_if(d, colname_pattern=vars(everything()), data_pattern=function(x){T}, condition=c("or", "and")[1] )
}
