% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_gender_cols_df}
\alias{select_gender_cols_df}
\title{A Function}
\usage{
select_gender_cols_df(df, output = c("df", "names", "dfnewnames"))
}
\description{
This function allows you to
}
\examples{
()
}
