% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_matches}
\alias{select_matches}
\title{Samantha Rhoads's function to extract or keep or select elements of a list, dataframe, or vector with a regular expression. Invert means keep everything other than the pattern. Default pattern is everything}
\usage{
select_matches(x, pat = ".*", invert = F, ignore.case = F)
}
\description{
Samantha Rhoads's function to extract or keep or select elements of a list, dataframe, or vector with a regular expression. Invert means keep everything other than the pattern. Default pattern is everything
}
\examples{
x <- as.list(iris); select_matches(x, pat="Petal"); v <- iris$Species; select_matches(v, pat="versi")
select_matches(x, pat=".*", invert=F, ignore.case=F)
}
