% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_name_race_gender_cols}
\alias{select_name_race_gender_cols}
\title{A Function}
\usage{
select_name_race_gender_cols(dat, type = c("list", "df"),
  output = NULL)

select_name_race_gender_cols(dat, type = c("list", "df"),
  output = NULL)
}
\description{
This function allows you to

This function allows you to
}
\examples{
()
()
}
