% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_not_race_gender_cols}
\alias{select_not_race_gender_cols}
\title{A Function}
\usage{
select_not_race_gender_cols(mylist)
}
\description{
This function allows you to
}
\examples{
select_not_race_gender_cols()
}
