% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_nrg_cols_list}
\alias{select_nrg_cols_list}
\title{A Function}
\usage{
select_nrg_cols_list(mylist, output = c("list", "names"))

select_nrg_cols_list(mylist, output = c("list", "names"))
}
\description{
This function allows you to

This function allows you to
}
\examples{
select_nrg_cols_list()
()
}
