% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_race_cols_list}
\alias{select_race_cols_list}
\title{A Function}
\usage{
select_race_cols_list(mylist, output = c("list", "names"))
}
\description{
This function allows you to
}
\examples{
select_race_cols_list()
}
