% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_unique_columns}
\alias{select_unique_columns}
\title{A function to select the unique columns in a dataframe}
\usage{
select_unique_columns(d)
}
\description{
A function to select the unique columns in a dataframe
}
\examples{
select_unique_columns(d)
}
