% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_vec}
\alias{select_vec}
\title{Samantha Rhoads's function to...}
\usage{
select_vec(v, pattern, ignore.case = T, everything = F)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
select_vec()
}
