% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_vec2}
\alias{select_vec2}
\title{Samantha Rhoads's function to extract or keep or select elements of a VECTOR. Invert means keep everything other than the pattern. Default pattern is everything}
\usage{
select_vec2(v, pattern = ".*", ignore.case = T, everything = F, invert = F)
}
\description{
Samantha Rhoads's function to extract or keep or select elements of a VECTOR. Invert means keep everything other than the pattern. Default pattern is everything
}
\examples{
select_matches(x, pat=".*", invert=F, ignore.case=F)
}
