% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_names_skip_rows_until_match_concat}
\alias{set_names_skip_rows_until_match_concat}
\title{Samantha Rhoads's function to set names to a specific row and concatenate the contents above into the names}
\usage{
set_names_skip_rows_until_match_concat(d, example_colname = "Employee ID")
}
\description{
Samantha Rhoads's function to set names to a specific row and concatenate the contents above into the names
}
\examples{
set_names_skip_rows_until_match_concat(d, example_colname="Employee ID")
}
