% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_names_skip_rows_until_match_loop}
\alias{set_names_skip_rows_until_match_loop}
\title{Samantha Rhoads's function to assign column names by skipping the appropriate number of rows in a dataframe before the actual column names. Uses a reference name.}
\usage{
set_names_skip_rows_until_match_loop(
  d,
  patterns = c("census_code", "census_title", "occp_code"),
  exact = F,
  check_n_rows = 30,
  doEvenIfColnameIsAlreadyIt = F,
  verbose = F
)
}
\description{
Samantha Rhoads's function to assign column names by skipping the appropriate number of rows in a dataframe before the actual column names. Uses a reference name.
}
\examples{
set_names_skip_rows_until_match_loop(d, patterns=c('census_code','census_title', 'occp_code'), exact=F, check_n_rows=30, doEvenIfColnameIsAlreadyIt=F)
}
