% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setdiff_}
\alias{setdiff_}
\title{Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from}
\usage{
<<<<<<< HEAD
setdiff_(
  x,
  y,
  returnWhichOnly = T,
  printWhichOnly = F,
  sortItems = F,
  return_as_tibble = F
)
=======
setdiff_(x, y, returnWhichOnly = T, printWhichOnly = F, sortItems = F)
>>>>>>> f405f1a7877bb7e35081dff4731c8a0360ba368e
}
\description{
Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from
}
\examples{
setdiff_(x, y, printWhichOnly=F)
}
