% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sort_str_by_alpha}
\alias{sort_str_by_alpha}
\title{Srhoads wrote this to allow you to...}
\usage{
sort_str_by_alpha(v, desc = F)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
sort_str_by_alpha()
}
