% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sort_vec_by_nchar}
\alias{sort_vec_by_nchar}
\title{Srhoads wrote this to allow you to sort a vector by number of characters}
\usage{
sort_vec_by_nchar(v, desc = T)
}
\description{
Srhoads wrote this to allow you to sort a vector by number of characters
}
\examples{
sort_str_by_nchar(v, desc = T)
}
