% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_before_capital}
\alias{split_before_capital}
\title{Samantha Rhoads's function to split a string between a lowercase and capital letter, as long as the letter after the uppercase one is a lowercase letter}
\usage{
split_before_capital(x, sep = " ", loosly = F)
}
\description{
Srhoads wrote this to allow you to split a string between a lowercase and capital letter, as long as the letter after the uppercase one is a lowercase letter (loosly means you don't need the letter after the capital letter to be lowercase)
}
\examples{
split_before_capital(x, sep=" ", loosly=F)
}
