% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_by_index}
\alias{split_by_index}
\title{Samantha Rhoads's function to split a dataframe by specified row indeces (numbers)}
\usage{
split_by_index(d, index = c(5, 10, 15))
}
\description{
Samantha Rhoads's function to split a dataframe by specified row indeces (numbers)
}
\examples{
split_by_index(d, index=c(5, 10, 15))
}
