% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_original_rdas}
\alias{split_original_rdas}
\title{A function}
\usage{
split_original_rdas(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  newdir = NULL,
  subsets = NULL,
  outpath = "~/",
  by,
  extra = NULL,
  filename_prefix,
  verbose = F
)
}
\description{
A function
}
\examples{
split_original_rdas()
}
