% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statetoabb}
\alias{statetoabb}
\title{Samantha Rhoads's function to...}
\usage{
statetoabb(v, state_names_abbs_df = get_states(return_which = "both"))
}
\description{
Samantha Rhoads's function to...
}
\examples{
statetoabb(v)
}
