% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{str_extract_money}
\alias{str_extract_money}
\title{Srhoads wrote this to allow you to...}
\usage{
str_extract_money(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
str_extract_money()
}
