% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{str_extract_zip}
\alias{str_extract_zip}
\title{Samantha Rhoads's function to...}
\usage{
str_extract_zip(v, concat = T, collapse = "; ")
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
str_extract_zip()
}
