% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strip_accents}
\alias{strip_accents}
\title{Samantha Rhoads's function to strip accents from characters}
\usage{
strip_accents(v, id = "Latin-ASCII")
}
\description{
Samantha Rhoads's function to strip accents from characters
}
\examples{
strip_accents(v, id="Latin-ASCII")
}
