% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{strip_num_trimws}
\alias{strip_num_trimws}
\title{A function}
\usage{
strip_num_trimws(v)

strip_num_trimws(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
strip_num_trimws()
strip_num_trimws()
}
