% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_factor}
\alias{summary_factor}
\title{A function}
\usage{
summary_factor(x, maxsum = 7)

summary_factor(x, maxsum = 7)
}
\description{
This function allows you to

Srhoads wrote this to allow you to...
}
\examples{
summary_factor()
summary_factor()
}
