% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_factor_2deep}
\alias{summary_factor_2deep}
\title{Srhoads wrote this to allow you to...}
\usage{
summary_factor_2deep(x, maxsum = 7)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
summary_factor_2deep()
}
