% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumrys}
\alias{sumrys}
\title{Samantha Rhoads's function to give length and sumry in one go}
\usage{
sumrys(x, n = 7)
}
\description{
Srhoads wrote this to allow you to get the length of the object and sumry of the object in one call (see `sumry()` for details on that type of summary)
}
\examples{
sumrys(x, n=7)
}
