% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sysdate}
\alias{sysdate}
\title{Samantha Rhoads's function to return today's date in yyyymmdd format (no punctuation)}
\usage{
sysdate()
}
\description{
Samantha Rhoads's function to return today's date in yyyymmdd format (no punctuation)
}
\examples{
sysdate()
}
