% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sysdateMDY}
\alias{sysdateMDY}
\title{Samantha Rhoads's function to return today's date in mmddyyyy format (no punctuation)}
\usage{
sysdateMDY()
}
\description{
Samantha Rhoads's function to return today's date in mmddyyyy format (no punctuation)
}
\examples{
sysdateMDY()
}
