% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{system_open}
\alias{system_open}
\title{Samantha Rhoads's function to tell the shell to open a file in whatever its default program is}
\usage{
system_open(paths)
}
\description{
Samantha Rhoads's function to tell the shell to open a file in whatever its default program is
}
\examples{
system_open(paths)
}
