% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tokdtmfunc}
\alias{tokdtmfunc}
\title{A Function}
\usage{
tokdtmfunc(df = fgsam, xv = "firstname", yv = "gender", nmin = 2,
  nmax = 3, term_count_min = 2)
}
\description{
This function allows you to
}
\examples{
()
}
