% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimpunct_}
\alias{trimpunct_}
\title{Samantha Rhoads's function to...}
\usage{
trimpunct_(v, removenegativesymbol = F, beginning = T, end = T)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
trimpunct_()
}
