% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{trimspace}
\alias{trimspace}
\title{A Function}
\usage{
trimspace(vec)

trimspace(vec)
}
\description{
This function allows you to

This function allows you to
}
\examples{
trimspace()
trimspace()
}
