% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{trimws_}
\alias{trimws_}
\title{A function}
\usage{
trimws_(v, which = "both", doublespace = T)

trimws_(v, which = "both", doublespace = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
trimws_()
trimws_()
}
