% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{trimws_df}
\alias{trimws_df}
\title{A function}
\usage{
trimws_df(x, which = "both", doublespace = T)

trimws_df(x, which = "both", doublespace = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
trimws_df()
trimws_df()
}
