% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_combine}
\alias{try_combine}
\title{A Function}
\usage{
try_combine(x)

try_combine(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_combine()
try_combine()
}
