% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_data_frame}
\alias{try_data_frame}
\title{A Function}
\usage{
try_data_frame(x)

try_data_frame(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_data_frame()
try_data_frame()
}
