% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_read_csv}
\alias{try_read_csv}
\title{A function to a read csv file}
\usage{
try_read_csv(file)

try_read_csv(file)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_read_csv()
try_read_csv()
}
