% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_read_dat}
\alias{try_read_dat}
\title{A function}
\usage{
try_read_dat(f)

try_read_dat(f)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_read_dat()
try_read_dat()
}
