% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{try_read_excel_somesheets}
\alias{try_read_excel_somesheets}
\title{Samantha Rhoads's function to trycatch reading excel data from a file or list of files}
\usage{
try_read_excel_somesheets(
  fns = NULL,
  keepshtvec = NULL,
  na = c("NA", "None", "N/A", "-", ""),
  col_types = "text",
  skip = 0
)
}
\description{
Samantha Rhoads's function to trycatch reading excel data from a file or list of files
}
\examples{
try_read_excel_somesheets(fns = NULL, keepshtvec = NULL, na = c("NA", "None", "N/A", "-", ""), col_types = "text", skip = 0)
}
