% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_read_file}
\alias{try_read_file}
\title{A Function}
\usage{
try_read_file(file, bindsheets = F)

try_read_file(file, bindsheets = F)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_read_file()
try_read_file()
}
