% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_read_table}
\alias{try_read_table}
\title{A Function}
\usage{
try_read_table(f)

try_read_table(f)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_read_table()
try_read_table()
}
