% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{try_unzip}
\alias{try_unzip}
\title{A function}
\usage{
try_unzip(zipfile)

try_unzip(zipfile)
}
\description{
This function allows you to

This function allows you to
}
\examples{
try_unzip()
try_unzip()
}
