% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{type_convert_exclude}
\alias{type_convert_exclude}
\title{Samantha Rhoads's function to convert column types automatically, without unncessary console messages, excluding variables named in the exclude_colname argument}
\usage{
type_convert_exclude(d, exclude_colname = "eeid")
}
\description{
Samantha Rhoads's function to convert column types automatically, without unncessary console messages, excluding variables named in the exclude_colname argument
}
\examples{
type_convert_exclude(d, exclude_colname="eeid")
}
