% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unique_sep_sort}
\alias{unique_sep_sort}
\title{Samantha Rhoads's function to}
\usage{
unique_sep_sort(v, sep = "; ")
}
\description{
Samantha Rhoads's function to
}
\examples{
unique_sep_sort(v, sep = "; ")
}
