% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unite_all}
\alias{unite_all}
\title{Samantha Rhoads's function to...}
\usage{
unite_all(
  d,
  clean = T,
  remove = F,
  newcol = "unite_all_column",
  onlynewcol = F,
  sep = "; "
)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
unite_all()
}
