% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unite_at}
\alias{unite_at}
\title{Srhoads wrote this to allow you to...}
\usage{
unite_at(
  d,
  fun = newcol,
  clean = T,
  remove = F,
  newcol = "unite_all_column",
  onlynewcol = F,
  sep = "; "
)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
unite_at(d, fun = newcol, clean = T, remove = F, newcol = "unite_all_column", onlynewcol = F, sep = "; ")
}
