% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unload_pkg}
\alias{unload_pkg}
\title{Samantha Rhoads's function to unload packages}
\usage{
unload_pkg(pkgs)
}
\description{
Samantha Rhoads's function to unload packages
}
\examples{
unload_pkg(pkgs)
}
