% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unzip_dir}
\alias{unzip_dir}
\title{A function to unzip a folder/directory}
\usage{
unzip_dir(zipfile, outdir = "unzip")
}
\description{
A function to unzip a folder/directory
}
\examples{
unzip_dir(zipfile, outdir="unzip")
}
