% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{windows2macPath}
\alias{windows2macPath}
\title{Samantha Rhoads's function to convert a windows path to a MAC path (specifically for the Y drive aapdata Volume mount)}
\usage{
windows2macPath(
 
    path = "Y:\\\\AA_Secured\\\\CVS\\\\2020\\\\2020 Plan Prep\\\\Aetna\\\\PDF Files for Sam\\\\Narrative Information"
)
}
\description{
Samantha Rhoads's function to convert a windows path to a MAC path (specifically for the Y drive aapdata Volume mount)
}
\examples{
windows2macPath(path)
}
