% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{word_to_front}
\alias{word_to_front}
\title{A function that brings a desired word to the front of a string}
\usage{
word_to_front(wrd, v)
}
\description{
This function brings a desired word to the front of a string (reorders string itself)
}
\examples{
word_to_front(wrd, v)
}
