% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_namesplit}
\alias{write_namesplit}
\title{A function}
\usage{
write_namesplit(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  newdir = NULL,
  extension = "feather",
  subsets = 2,
  subsubsets = 2,
  extra = NULL,
  outpath = "~/",
  filename_prefix = ""
)
}
\description{
This function allows you to
}
\examples{
write_namesplit()
}
