% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writexl_with_formulas}
\alias{writexl_with_formulas}
\title{Samantha Rhoads's function to write to Excel just like writexl::write_xlsx (same args) but you can specify Excel formulas to be written based on your current columns}
\usage{
writexl_with_formulas(
  x,
  path = tempfile(fileext = ".xlsx"),
  formula = "=total - white",
  newcol = "minority",
  col_names = T,
  format_headers = T,
  use_zip64 = F,
  write = F
)
}
\description{
Samantha Rhoads's function to write to Excel just like writexl::write_xlsx (same args) but you can specify Excel formulas to be written based on your current columns
}
\examples{
writexl_with_formulas(x, path = tempfile(fileext = ".xlsx"), formula="=total - white", newcol="minority", col_names = T, format_headers = T, use_zip64 = F, write=F)
}
