% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zip5}
\alias{zip5}
\title{A function kinda like `dplyr`'s `select()` but works on lists}
\usage{
zip5(text.var, trim = !extract, clean = TRUE,
  pattern = "(?<!\\\\d)\\\\d{5}(?!\\\\d)", replacement = "",
  extract = TRUE, dictionary = getOption("regex.library"), ...)
}
\description{
This function is kinda like `dplyr`'s `select()` but works on selecting stuff from lists
}
\examples{
zip5()
}
