% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zipcode5}
\alias{zipcode5}
\title{A function that gets 5 digit zip codes; it extracts relevant first 5 digits from zip code when it's formatted like 11111-1111}
\usage{
zipcode5(v)
}
\description{
A function that gets 5 digit zip codes; it extracts relevant first 5 digits from zip code when it's formatted like 11111-1111
}
\examples{
{v <- c('924832sjkhf',' 20439-22912', '93488-9293', '23984_93284', '39842-3949 32432', '328rjd2 nh22 22242 ')}
zipcode5(v)
}
