% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_recode}
\alias{clean_recode}
\title{A function}
\usage{
clean_recode(
  df,
  scrub = c("once", "double"),
  extrarace = NULL,
  extragender = NULL
)
}
\description{
A function
}
\examples{
clean_recode()
}
