% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{closest}
\alias{closest}
\title{Jason originally wrote this to allow you to find the closest date in your piece of data to a given date you choose, like mathcing them up. Now it's Samantha Rhoads's function to find the closest date}
\usage{
closest(x, y, type = NA)
}
\description{
Jason originally wrote this to allow you to find the closest date in your piece of data to a given date you choose, like mathcing them up. Now it's Samantha Rhoads's function to find the closest date
}
\examples{
closest(x, y, type=NA)
}
