% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_tiered_vec}
\alias{collapse_tiered_vec}
\title{Samantha Rhoads's function to...}
\usage{
collapse_tiered_vec(v, collapse = "; ")
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
collapse_tiered_vec()
}
