% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{datatable_}
\alias{datatable_}
\title{A function}
\usage{
datatable_(d)

datatable_(d)
}
\description{
This function allows you to

This function allows you to
}
\examples{
datatable_()
datatable_()
}
