% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df_get_preferred_column}
\alias{df_get_preferred_column}
\title{Samantha Rhoads's function to get relevant, preferred, desired column from dataframe on a loop}
\usage{
df_get_preferred_column(
  df,
  patterns = c("DateOpened", "Date.*Opened"),
  ignore.case = T,
  fillmissingwith = NA,
  returnNameOnly = F,
  exactEnd = F,
  exactStart = F,
  verbose = F
)
}
\description{
Samantha Rhoads's function to get relevant, preferred, desired column from dataframe on a loop
}
\examples{
df_get_preferred_column(df, patterns=c('DateOpened', 'Date.*Opened'), ignore.case=T, fillmissingwith=NA, returnNameOnly=F, exactEnd=F, exactStart=F)
}
