% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dfsampler}
\alias{dfsampler}
\title{A function that makes sample data frames with name, race, gender type columns}
\usage{
dfsampler(which = "long", tibble = F)
}
\description{
This function allows you to make sample dataframes!
}
\examples{
dfsampler(which='long', tibble=F)
}
