% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{dply_nonnumeric}
\alias{dply_nonnumeric}
\title{A Function}
\usage{
dply_nonnumeric(x, fun)

dply_nonnumeric(x, fun)
}
\description{
This function allows you to

This function allows you to
}
\examples{
dply_nonnumeric()
dply_nonnumeric()
}
