% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excelToDateIf5DigitStr}
\alias{excelToDateIf5DigitStr}
\title{Samantha Rhoads's function to...}
\usage{
excelToDateIf5DigitStr(v)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
excelToDateIf5DigitStr()
}
