% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_join_first_last_name}
\alias{gather_join_first_last_name}
\title{A Function}
\usage{
gather_join_first_last_name(df, seq = c("first last", "last,first"))
}
\description{
This function allows you to
}
\examples{
()
}
