% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMostRecentFiles}
\alias{getMostRecentFiles}
\title{Samantha Rhoads's function to return the most recent versions of files based on a given pattern of characters of its name}
\usage{
getMostRecentFiles(
  path = ".",
  desc = T,
  verbose = F,
  pattern = NULL,
  all.files = F,
  full.names = T,
  recursive = T,
  ignore.case = F,
  include.dirs = F,
  no.. = F,
  printfilenames = F
)
}
\description{
Samantha Rhoads's function to return the most recent versions of files based on a given pattern of characters of its name
}
\examples{
getMostRecentFiles(path=".", desc=T, verbose=F, pattern=NULL, all.files=F, full.names=T, recursive=T, ignore.case=F, include.dirs=F, no..=F, printfilenames=F)
}
