% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grep_ic}
\alias{grep_ic}
\title{Samantha Rhoads's function to...}
\usage{
grep_ic(pattern, x, ignore.case = T, perl = FALSE, value = FALSE,
  fixed = FALSE, useBytes = FALSE, invert = FALSE)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
grep_ic()
}
