% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grepl_ic}
\alias{grepl_ic}
\title{Samantha Rhoads's function to...}
\usage{
grepl_ic(
  pattern,
  x,
  ignore.case = T,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
grepl_ic()
}
