% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{is.factorchar}
\alias{is.factorchar}
\title{A function to ask if it's a character or factor logical}
\usage{
is.factorchar(x)

is.factorchar(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
is.factorchar()
is.factorchar()
}
