% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.nanull}
\alias{is.nanull}
\title{Srhoads wrote this to allow you to check if all of a variable is NA or NULL (edited 20200226)}
\usage{
is.nanull(x)
}
\description{
Srhoads wrote this to allow you to check if all of a variable is NA or NULL (edited 20200226)
}
\examples{
is.nanull(x)
}
