% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_firstlastname}
\alias{join_firstlastname}
\title{A function}
\usage{
join_firstlastname(
  df,
  firstname,
  lastname,
  seq = c("first last", "last,first")
)
}
\description{
This function allows you to
}
\examples{
join_firstlastname()
}
