% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{lastname_after_space_when_nocomma}
\alias{lastname_after_space_when_nocomma}
\title{A Function}
\usage{
lastname_after_space_when_nocomma(v)

lastname_after_space_when_nocomma(v)
}
\description{
This function allows you to

This function allows you to
}
\examples{
lastname_after_space_when_nocomma()
lastname_after_space_when_nocomma()
}
