% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lastname_after_space_when_nocomma}
\alias{lastname_after_space_when_nocomma}
\title{A function}
\usage{
lastname_after_space_when_nocomma(v)
}
\description{
This function allows you to
}
\examples{
lastname_after_space_when_nocomma()
}
