% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{list_files}
\alias{list_files}
\title{A Function}
\usage{
list_files(p, pattern = NULL, recursive = T, full.names = T)

list_files(p, pattern = NULL, recursive = T, full.names = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
list_files()
list_files()
}
