% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{not_all_na}
\alias{not_all_na}
\title{A Function}
\usage{
not_all_na(x)

not_all_na(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
not_all_na()
not_all_na()
}
