% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{pkg}
\alias{pkg}
\title{A Function}
\usage{
pkg(package1, ...)

pkg(package1, ...)
}
\description{
This function allows you to

This function allows you to
}
\examples{
pkg()
pkg()
}
