% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pkg2}
\alias{pkg2}
\title{A function to load and/or install package first (holistic)}
\usage{
pkg2(package1 = NULL, ..., pipes = T)
}
\description{
This function allows you to load and/or install package first (holistic)
}
\examples{
pkg()
}
