% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_excel_somesheets}
\alias{read_excel_somesheets}
\title{A function}
\usage{
read_excel_somesheets(
  fns = NULL,
  keepshtvec = NULL,
  na = c("NA", "None", "N/A", "-", ""),
  col_types = "text",
  skip = 0
)
}
\description{
A function
}
\examples{
read_excel_somesheets(fns=NULL, keepshtvec=NULL, na=c("NA", "None", "N/A", "-", ""), col_types='text', skip=0)
}
