% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_excels}
\alias{read_excels}
\title{A function to read excel files}
\usage{
read_excels(filelist, bindsheets = F, joinsheets = F, bindrows = F,
  simplif = T)
}
\description{
This function allows you to
}
\examples{
read_excels(filelist, bindsheets=F, joinsheets=F, bindrows=F, simplif=T)
}
