% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_excels}
\alias{read_excels}
\title{A function}
\usage{
read_excels(
  filelist,
  bindsheets = F,
  bindrows = F,
  simplif = F,
  col_types = "text"
)
}
\description{
A function
}
\examples{
read_excels(filelist, bindsheets = F, bindrows = F, simplif = F, col_types = "text")
}
