% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_excels}
\alias{read_excels}
\title{A function to read excel files}
\usage{
read_excels(
  filelist,
  bindsheets = F,
  bindrows = F,
  simplif = T,
  col_types = "text"
)

read_excels(
  filelist,
  bindsheets = F,
  bindrows = F,
  simplif = T,
  col_types = "text"
)

read_excels(
  filelist,
  bindsheets = F,
  bindrows = F,
  simplif = T,
  col_types = "text"
)
}
\description{
This function allows you to

#' A function
#'
#' This function allows you to 
#' @export
#' @examples
#' df_paste_2()
df_paste_2 <- function(df){
  vector_paste(c("'____'", substitute(df), "'____'"))
  df_paste(df)
}

This function allows you to
}
\details{
#' A function
#'
#' This function allows you to 
#' @export
#' @examples
#' df_paste_lod()
df_paste_lod <- function(lod){
  lapply(seq_along(lod), 
         function(y, n, i) {vector_paste(n[[i]]); df_paste(y[[i]])}, 
         y=lod, n=names(lod))
}

A function

This function allows you to
}
\examples{
read_excels()
read_excels()
read_excels()
}
