% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{read_files}
\alias{read_files}
\title{A function}
\usage{
read_files(filelist, bindsheets = F, bindrows = F, simplif = T)

read_files(filelist, bindsheets = F, bindrows = F, simplif = T)
}
\description{
This function allows you to

This function allows you to
}
\examples{
read_files()
read_files()
}
