% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files}
\alias{read_files}
\title{A function}
\usage{
read_files(filelist, bindsheets = F, bindrows = F, simplif = T)
}
\description{
A function
}
\examples{
read_files(filelist, bindsheets=F, bindrows=F, simplif=T)
}
