% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_files_to_feather}
\alias{read_files_to_feather}
\title{A Function}
\usage{
read_files_to_feather(filelist, path = "~/",
  filename_prefix = "nrg_data")
}
\description{
This function allows you to
}
\examples{
read_files_to_feather()
}
