% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_merge_feathers}
\alias{read_merge_feathers}
\title{A Function}
\usage{
read_merge_feathers(filelist = NULL, inpath = NULL, pattern = NULL,
  recode_extra_na = F)
}
\description{
This function allows you to
}
\examples{
()
}
