% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_merge_write_feathers}
\alias{read_merge_write_feathers}
\title{A function}
\usage{
read_merge_write_feathers(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  newdir = NULL,
  outpath = "~/",
  filename_prefix = "DEFAULT"
)
}
\description{
A function
}
\examples{
read_merge_write_feathers()
}
