% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_rdas}
\alias{read_rdas}
\title{A function}
\usage{
read_rdas(
  filelist = NULL,
  inpath = NULL,
  pattern = NULL,
  bind = F,
  recode_extra_na = F
)
}
\description{
A function
}
\examples{
read_rdas()
}
