% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_rdas}
\alias{read_rdas}
\title{A Function}
\usage{
read_rdas(filelist = NULL, inpath = NULL, pattern = NULL)

read_rdas(filelist = NULL, inpath = NULL, pattern = NULL)
}
\description{
This function allows you to

This function allows you to
}
\examples{
()
()
}
