% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_ydrive_clean_write}
\alias{read_ydrive_clean_write}
\title{A Function}
\usage{
read_ydrive_clean_write(filenames, csv = F, xlsx = F, xls = F,
  outpath = "data/original/")
}
\description{
This function allows you to
}
\examples{
()
}
