% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readexcel}
\alias{readexcel}
\title{A function just like read_excel but better bc it does all the sheets!}
\usage{
readexcel(file, bindsheets = F, skip = 0, col_types = "text", simplify = T)
}
\description{
A function just like read_excel but better bc it does all the sheets!
}
\examples{
readexcel(file, bindsheets=F, skip=0, col_types='text', simplify=T)
}
