% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_from_list}
\alias{recode_from_list}
\title{A function that acts like `recode()` but works on a list}
\usage{
recode_from_list(
  v,
  recode_list = list(`NA` = c("na", "not applicable"), `TRUE` = c("true", "tru"))
)
}
\description{
A function that acts like `recode()` but works on a list
}
\examples{
recode_from_list(v, recode_list = list('NA' = c('na', 'not applicable'), 'TRUE' = c('true', 'tru')))
}
