% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{recode_na}
\alias{recode_na}
\title{A function}
\usage{
recode_na(x)

recode_na(x)
}
\description{
This function allows you to

This function allows you to
}
\examples{
recode_na()
()
}
