% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_na_vec}
\alias{recode_na_vec}
\title{A function}
\usage{
recode_na_vec(
  vec,
  recode_list = recode_na_list,
  extra = NULL,
  recode_na_getridofstrregex = NULL
)
}
\description{
A function
}
\examples{
recode_na_vec()
}
