% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_race_j}
\alias{recode_race_j}
\title{A Function}
\usage{
recode_race_j(vec, recode_list = race_list, extra = c("nothing",
  "multirace", "NA"))
}
\arguments{
\item{recode_list}{List of gender categories and codes Defaults to a built-in comprehensive list.}

\item{extra}{Extra categories that weren't able to be recoded. extra = c("nothing", "other", "NA"). Defaults to "nothing".}
}
\description{
This function allows you to recode race
}
\examples{
recode_race_j(vec, recode_list = race_list, extra = c("nothing", "multirace", "NA"))
}
