% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode_state}
\alias{recode_state}
\title{Samantha Rhoads's function to recode states (even from state fips codes)}
\usage{
recode_state(v, abb = T, to_fips = F)
}
\description{
Samantha Rhoads's function to recode states (even from state fips codes)
}
\examples{
recode_state(v, abb=T, to_fips=F)
}
