% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setdiff_}
\alias{setdiff_}
\title{Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from}
\usage{
setdiff_(x, y, returnWhichOnly = T, printWhichOnly = F, sortItems = F)
}
\description{
Samantha Rhoads's function to do `setdiff()` from both sides and print which input the different strings come from
}
\examples{
setdiff_(x, y, printWhichOnly=F)
}
