% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{state2abb_or_abb2state}
\alias{state2abb_or_abb2state}
\title{Samantha Rhoads's function to...}
\usage{
state2abb_or_abb2state(v, abb = F)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
state2abb_or_abb2state()
}
