% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strip_punct}
\alias{strip_punct}
\title{Samantha Rhoads's function to...}
\usage{
strip_punct(v, replacewithspace = T, replacewith = NULL,
  onlyends = F)
}
\description{
Srhoads wrote this to allow you to...
}
\examples{
strip_punct()
}
