% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_all_paste0}
\alias{summarize_all_paste0}
\title{Samantha Rhoads's function to summarize_all remaining columns in a grouped df by pasting their unique values}
\usage{
summarize_all_paste0(
  .tbl,
  .funs,
  ...,
  collapse = "; ",
  unique_sep_sort_str = T,
  unique_sep_str = T,
  sort_str = T,
<<<<<<< HEAD
  recode_na_vals = c("", "NA", "NaN"),
  strip_na = T,
=======
  recode_na_vals = c("", "NA"),
>>>>>>> f405f1a7877bb7e35081dff4731c8a0360ba368e
  ungroup = F
)
}
\description{
Samantha Rhoads's function to summarize_all remaining columns in a grouped df by pasting their unique values
}
\examples{
summarize_all_paste0(.tbl, .funs, ..., collapse=", ", unique_sep_sort_str=T, recode_na_vals=c("", "NA"), ungroup=F)
}
