% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimws_}
\alias{trimws_}
\title{This function allows you to trim whitespace but also remove double spaces}
\usage{
trimws_(v, which = "both", doublespace = T)
}
\description{
This function allows you to trim whitespace but also remove double spaces
}
\examples{
trimws_(v, which='both', doublespace=T)
}
