% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{try_read_file}
\alias{try_read_file}
\title{A function to try reading any file from the read_file() function...tryCatch specific}
\usage{
try_read_file(file, bindsheets = F, joinsheets = F)
}
\description{
This function allows you to
}
\examples{
try_read_file(file, bindsheets=F, joinsheets=F)
}
