% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_generalizable.R
\name{unzip_dir}
\alias{unzip_dir}
\title{A function}
\usage{
unzip_dir(zipfile, outdir = "unzip")

unzip_dir(zipfile, outdir = "unzip")
}
\description{
This function allows you to

This function allows you to
}
\examples{
unzip_dir()
()
}
