% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writexl_open_formatted}
\alias{writexl_open_formatted}
\title{Samantha Rhoads's function to write to Excel with some common dataframe special formatting}
\usage{
writexl_open_formatted(
  df = NULL,
  filename = NULL,
  open_file = T,
  maxcolwidth = 50,
  colwidthplus = 0,
  freeze_after_col = c("^(EEID|eeid)$", 1)[2]
)
}
\description{
Samantha Rhoads's function to write to Excel with some common dataframe special formatting
}
\examples{
writexl_open_formatted(df=NULL, filename, open_file=T, max_colwidth=50, colwidthplus=0)
}
