% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writexl_open_formatted}
\alias{writexl_open_formatted}
\title{Samantha Rhoads's function to write to Excel with some common dataframe special formatting}
\usage{
writexl_open_formatted(
  x = NULL,
  filename = NULL,
  open_file = T,
  maxcolwidth = 50,
  colwidthplus = 0,
  freeze_after_col = c("^(EEID|eeid)$", 1)[2],
  autofilter = T,
  baseFontSize = 10,
  clean_colnames = T,
  colnames_toupper = F,
  force_as_tempfile = F,
  sheet_args = list(`1` = list(wrap_headers = T), `2` = list(wrap_all = F))
)
}
\description{
Samantha Rhoads's function to write to Excel with some common dataframe special formatting
}
\examples{
writexl_open_formatted(df=NULL, filename, open_file=T, max_colwidth=50, colwidthplus=0)
}
